integration_tests/test-requirements.txt
https://github.com/wazo-platform/wazo-test-helpers/archive/master.zip
kombu==4.6.11
pyhamcrest
pytest
python-consul==1.1.0
